requirements.txt
langchain>=0.3.26
langchain-core
langchain-community
langchain-google-genai
langchain-groq
langgraph
Pillow
python-dotenv
